#785 fix fast particle models
